First pass at supporting machine room switches
- Previously, all port configuration was based on wall plate number.  This
doesn't work in machine rooms and we had no way of changing interfaces on
those switches.  Now, if you put in a room number of 531 or 310, you are
prompted to enter the raw switch data instead off a wall plate number.
- This also includes the start of some sanity checking on ports.  It
should probably be moved out of the Console.pm and into Switch.pm at some point,
but I need to better understand the different scenarios under which this data
needs to be checked before doing so.