support for adding/deleting DNS aliases
- default dns_region is now 'all'.  When I get the split DNS support ready,
users will be able to pick one of 'internal', 'external', or 'all' for
dns_regions.
- created the udb-add-dns-alias and udb-del-dns-alias commands to allow
users to manage DNS aliases.  The add functions can either associate the
DNS alias with an existing DNS alias or create a brand new one.
- added functionality to get_ip_and_vlan subroutine, allowing you to specify
whether or not to include the dynamic notifications.  This allowed me to
generalize the function and use it for DNS aliases as well.