Re-wrote udb query to work on non-devices tables.
Given a number of rules of the form:

(table.)?field=value

udb-query returns the hostnames of machines which match all rules.
Thus, it still only works on tables with a device_name field.

If a table name is not supplied, it works on the devices table.

If you want the union instead of the intersection, run multiple queries
and pipe the output to uniq.

$(udb query computers.os_type=debian; \
udb query usage=other;) | \
uniq

This will break a format that relied on apostrophes. I didn't recognize
that syntax and don't know of anything using it.

Rules should probably be checked for validity before being processed.
-POLM