Support for adding and deleting interfaces to/from a device
- ability to add interfaces associated with a new ip address
and one that will be used as a bonded interface, i.e. multiple
interfaces on a single device that act like one interface
- added two db triggers so extraneous data isn't left around
in db after devices or interfaces are removed
- cleaned up dns alias deletion, as extra code is no longer
needed with additional db triggers
- fixed bug in choose_interface function, which caused it to
be a no-op
- fixed room detection for 431A
- moved dns_insert to Net.pm
- moved verify_ip_or_vlan to Net.pm
- moved verify_mac to Net.pm
- fixed DNS templates, to match multiple interfaces and ensure
only one MX record is entered for each host.  This requires that
A records and MX records be separate in DNS files now