Support for staged additions, modifications, and deletions
- Add support for setting and querying timestamps.  Introduced
a build_log table, that tracks last successful build timestamps
for various tables.  This way, we can trigger certain build
activities based on changes to various tables.
- Mark devices for deletion instead of instantly removing them
from the database.
- Add checks for various actions (i.e. move) that ensures people
can't operate on devices with deleted status.
- Introduced state table and in particular a build_lock entry.
This will ensure that multiple builds do not occur at the same
time.  In case there is a bug, added a udb-reset-build-lock
script.  This command is not referenced by the udb wrapper, so
that people can't run it by mistake.  It "should" not be needed,
but wanted it there just in case.
- Add logic to extract a udbuser/admin keytab entry.  This is
the entry actually used by kadmin to add and delete host
pricipals.  Unfortunately, this requires that a users admin
credentials must be autheticated during each and every build.
- Quieted down some of the daemon restarts, when not running in
verbose mode.