[dns] use unix time for serial numbers
Instead of a sequence in the Postgres DB, just use UNIX time. It does
everything DNS zone serial number needs to, is compatible with the old
convention, and represents something meaningful.

This was motivated by me seeing zone files with (I kid you not) sequence
numbers 12345 and wondering where the heck *that* was hard-coded.